[wip] add MobileRefinery Trait
Attempts creating a new C# Trait class and adding that to a unit. However, the TraitInfo for this new Trait currently can't be found when you run the game, causing a crash. I'm sure it has something to do with getting the engine to import the directory properly, but I haven't figured out yet where I'm messing this up.