fix MobileRefinery Trait and sprite crashes for Extractor
However, the Extractor now renders two sprites on top of itself: the MobileRefinery sprite, which is a looping Cyborg image crouching down, and its regular infantry sprite, which animates properly and runs around the world.

This makes total sense of course; we created a new Trait that is a clone of a building, and buildings expect to have a sprite rendering of themselves. Need to rip out the sprite code from the MobileRefinery trait so that "anything" can be a MobileRefinery and it doesn't automatically assume it needs a sprite; then we can tack MobileRefinery onto any unit or structure, and we have a more generic, less coupled Trait.