modify OpenRA engine files to allow docking within radius
Units with a bounding box as small as the infantry Harvesters that I'm using had trouble getting to the cell where a Refinery (even a custom MobileRefinery) said that Harvesters should go. As a result, they'd approach the Refinery and just stand there.

These changes address that, by allowing Harvesters to get "close enough" to a docking point by defining an acceptable radius for Harvesters to dock with the (Mobile)Refinery within the (Mobile)Refinery Trait. Specifying 0 or not specifying the key at all will cause the code to behave like it always did.

Used an absurd radius for now as a test of the system. I.e. radius of 5 did nothing; radius of 5000 makes the code execute as expected. Will have to fine-tune an actual proper radius setting for our specific units later.